GPD sync refactoring - part2
