--- django_api/etools_prp/apps/unicef/pgd_sync/process_gd_item.py
@@ -1,5 +1,0 @@
-from etools_prp.apps.core.models import Workspace
-
-
-def process_gd_item(item: dict, workspace: Workspace) -> bool:
-    return True
