+++ django_api/etools_prp/apps/unicef/sync/process_pd_item.py
@@ -1,30 +1,28 @@
 import logging
 
 from etools_prp.apps.core.models import Workspace
-from etools_prp.apps.unicef.ppd_sync.update_create_blueprint import update_create_blueprint
-from etools_prp.apps.unicef.ppd_sync.update_create_date_period import (
+from etools_prp.apps.unicef.sync.update_create_blueprint import update_create_blueprint
+from etools_prp.apps.unicef.sync.update_create_date_period import (
     update_create_qpr_n_hr_date_periods,
     update_create_sr_date_periods,
 )
-from etools_prp.apps.unicef.ppd_sync.update_create_disaggregation import update_create_disaggregations
-from etools_prp.apps.unicef.ppd_sync.update_create_expected_result import (
+from etools_prp.apps.unicef.sync.update_create_disaggregation import update_create_disaggregations
+from etools_prp.apps.unicef.sync.update_create_expected_result import (
     update_create_expected_result_llo,
     update_create_expected_result_rl,
 )
-from etools_prp.apps.unicef.ppd_sync.update_create_location import update_create_locations
-from etools_prp.apps.unicef.ppd_sync.update_create_partner import update_create_partner
-from etools_prp.apps.unicef.ppd_sync.update_create_pd import update_create_pd
-from etools_prp.apps.unicef.ppd_sync.update_create_person import (
+from etools_prp.apps.unicef.sync.update_create_location import update_create_locations
+from etools_prp.apps.unicef.sync.update_create_partner import update_create_partner
+from etools_prp.apps.unicef.sync.update_create_pd import update_create_pd
+from etools_prp.apps.unicef.sync.update_create_person import (
     update_create_agreement_auth_officers,
     update_create_focal_points,
     update_create_unicef_focal_points,
 )
-from etools_prp.apps.unicef.ppd_sync.update_create_reportable import update_create_reportable
-from etools_prp.apps.unicef.ppd_sync.update_create_reportable_location_goal import (
-    update_create_reportable_location_goals,
-)
-from etools_prp.apps.unicef.ppd_sync.update_create_section import update_create_sections
-from etools_prp.apps.unicef.ppd_sync.update_llos_and_reportables import update_llos_and_reportables
+from etools_prp.apps.unicef.sync.update_create_reportable import update_create_reportable
+from etools_prp.apps.unicef.sync.update_create_reportable_location_goal import update_create_reportable_location_goals
+from etools_prp.apps.unicef.sync.update_create_section import update_create_sections
+from etools_prp.apps.unicef.sync.update_llos_and_reportables import update_llos_and_reportables
 
 logger = logging.getLogger(__name__)
 
@@ -43,23 +41,23 @@ def process_pd_item(item: dict, workspace: Workspace) -> bool:
 
     if not item['start_date']:
         logger.warning("Start date is required - skipping!")
-        return False
+        return
 
     if not item['end_date']:
         logger.warning("End date is required - skipping!")
-        return False
+        return
 
     # Get partner
     item, partner = update_create_partner(item)
 
     if partner is None:
-        return False
+        return
 
     # Get PD
     item, pd = update_create_pd(item, workspace)
 
     if pd is None:
-        return False
+        return
 
     # Get Unicef Focal Points
     pd = update_create_unicef_focal_points(item['unicef_focal_points'], pd)
@@ -112,5 +110,3 @@ def process_pd_item(item: dict, workspace: Workspace) -> bool:
 
                 # Create Reportable Location Goals
                 update_create_reportable_location_goals(reportable, locations)
-
-    return True
