+++ django_api/etools_prp/apps/unicef/sync/update_create_partner.py
@@ -5,7 +5,7 @@ from rest_framework.exceptions import ValidationError
 
 from etools_prp.apps.partner.models import Partner
 from etools_prp.apps.partner.serializers import PMPPartnerSerializer
-from etools_prp.apps.unicef.ppd_sync.utils import process_model
+from etools_prp.apps.unicef.sync.utils import process_model
 
 logger = logging.getLogger(__name__)
 
@@ -14,12 +14,12 @@ def update_create_partner(item: Any) -> (Optional[Any], Optional[Partner]):
 
     # Skip entries without unicef_vendor_number
     if not item['partner_org']['unicef_vendor_number']:
-        logger.warning("No unicef_vendor_number - skipping!")
+        logger.warning("No unicef_vendor_number for document - skipping!")
         return item, None
 
     # Create/Assign Partner
     if not item['partner_org']['name']:
-        logger.warning("No partner name - skipping!")
+        logger.warning("No partner name for document - skipping!")
         return item, None
 
     item['partner_org']['external_id'] = item['partner_org'].get('id', '#')
